rename(#134): 패키지명 변경 event -> messaging
