+++ review/src/main/java/com/ticketing/review/application/service/EventService.java
@@ -1,7 +1,7 @@
 package com.ticketing.review.application.service;
 
-import com.ticketing.review.domain.event.ReviewDeleteErrorEvent;
-import com.ticketing.review.domain.event.ReviewTopic;
+import com.ticketing.review.domain.messaging.ReviewDeleteErrorEvent;
+import com.ticketing.review.domain.messaging.ReviewTopic;
 import com.ticketing.review.infrastructure.messaging.EventSerializer;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
